refactoring: event / annotation handling in graph panel broken out
