+++ public/app/features/annotations/event.ts
@@ -1,0 +1,10 @@
+
+export class AnnotationEvent {
+  dashboardId: number;
+  panelId: number;
+  time: any;
+  timeEnd: any;
+  isRegion: boolean;
+  title: string;
+  text: string;
+}
