+++ public/app/features/annotations/event_editor.ts
@@ -4,16 +4,7 @@ import _ from 'lodash';
 import moment from 'moment';
 import {coreModule} from 'app/core/core';
 import {MetricsPanelCtrl} from 'app/plugins/sdk';
-
-export class AnnotationEvent {
-  dashboardId: number;
-  panelId: number;
-  time: any;
-  timeEnd: any;
-  isRegion: boolean;
-  title: string;
-  text: string;
-}
+import {AnnotationEvent} from './event';
 
 export class EventEditorCtrl {
   panelCtrl: MetricsPanelCtrl;
@@ -73,61 +64,3 @@ export function eventEditor() {
 }
 
 coreModule.directive('eventEditor', eventEditor);
-
-export class EventManager {
-  event: AnnotationEvent;
-
-  constructor(private panelCtrl: MetricsPanelCtrl,
-              private elem,
-              private popoverSrv) {
-              }
-
-              editorClosed() {
-                console.log('editorClosed');
-                this.event = null;
-                this.panelCtrl.render();
-              }
-
-              updateTime(range) {
-                let newEvent = true;
-
-                if (this.event) {
-                  newEvent = false;
-                } else {
-                  // init new event
-                  this.event = new AnnotationEvent();
-                  this.event.dashboardId = this.panelCtrl.dashboard.id;
-                  this.event.panelId = this.panelCtrl.panel.id;
-                }
-
-                // update time
-                this.event.time = moment(range.from);
-                this.event.isRegion = false;
-                if (range.to) {
-                  this.event.timeEnd = moment(range.to);
-                  this.event.isRegion = true;
-                }
-
-                // newEvent means the editor is not visible
-                if (!newEvent) {
-                  this.panelCtrl.render();
-                  return;
-                }
-
-                this.popoverSrv.show({
-                  element: this.elem[0],
-                  classNames: 'drop-popover drop-popover--form',
-                  position: 'bottom center',
-                  openOn: null,
-                  template: '<event-editor panel-ctrl="panelCtrl" event="event" close="dismiss()"></event-editor>',
-                  onClose: this.editorClosed.bind(this),
-                  model: {
-                    event: this.event,
-                    panelCtrl: this.panelCtrl,
-                  },
-                });
-
-                this.panelCtrl.render();
-              }
-}
-
